fastapi-solution/tests/functional/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.8.4
elasticsearch==8.7.0
redis==4.5.5
python-dotenv==1.0.0
pydantic==1.10.9